created empty offers page screen
